requirements.txt
# Core Data Science
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

# Financial/Statistical Models
arch>=6.3.0
statsmodels>=0.14.0

# Visualization
plotly>=5.18.0
dash>=2.14.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Async/Networking
aiohttp>=3.9.0
requests>=2.31.0
websockets>=12.0

# Neo Blockchain
neo-mamba>=1.5.1
neo3-boa>=1.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
